adding rizzino_a07b.js
Still having problems with the VM.  Have updated some of the code but have still not been able to get my program working due to problems with the VM taking up much of my time I had for completing this assignment.  I am turning in what I have completed so far so something will be submitted on time.